duplicate code removed by defining a BaseLayout
