+++ test/layout/ConsultantLayout.spec.ts
@@ -1,27 +1,12 @@
 import { mount } from '@vue/test-utils';
-import { describe, it, expect, beforeEach, vi } from 'vitest';
+import { describe, it, expect, beforeEach } from 'vitest';
 import ConsultantLayout from '../../src/layout/ConsultantLayout.vue';
+import BaseLayout from '../../src/layout/BaseLayout.vue';
+import AppTopbar from '../../src/layout/AppTopbar.vue';
+import AppFooter from '../../src/layout/AppFooter.vue';
+import ConsultantMenu from '../../src/layout/ConsultantMenu.vue';
 import { createRouter, createWebHistory } from 'vue-router';
 import { createPinia } from 'pinia';
-
-vi.mock('@/layout/composables/layout', () => ({
-    useLayout: vi.fn(() => ({
-      layoutConfig: {
-        darkTheme: false,
-        menuMode: { value: 'static' },
-        ripple: { value: true },
-        fullscreen: { value: false },
-      },
-      layoutState: {
-        staticMenuDesktopInactive: { value: false },
-        overlayMenuActive: { value: false },
-        staticMenuMobileActive: { value: false },
-        menuHoverActive: { value: false },
-      },
-      isSidebarActive: false,
-      setFullscreen: vi.fn(),
-    })),
-  }));
 
 describe('ConsultantLayout.vue', () => {
   let router;
@@ -36,32 +21,23 @@ describe('ConsultantLayout.vue', () => {
     });
   });
 
-  it('should render all layout components', async () => {
+  it('should render BaseLayout with the correct slot components', async () => {
     const wrapper = mount(ConsultantLayout, {
       global: {
         plugins: [pinia, router],
       },
     });
 
-    expect(wrapper.findComponent({ name: 'AppTopbar' }).exists()).toBe(true);
-    expect(wrapper.findComponent({ name: 'AppFooter' }).exists()).toBe(true);
-    expect(wrapper.findComponent({ name: 'ConsultantMenu' }).exists()).toBe(true);
-    expect(wrapper.findComponent({ name: 'RouterView' }).exists()).toBe(true);
-  });
+    const baseLayout = wrapper.findComponent(BaseLayout);
+    expect(baseLayout.exists()).toBe(true);
 
-  it('should apply the correct layout class based on layoutConfig', async () => {
-    const wrapper = mount(ConsultantLayout, {
-      global: {
-        plugins: [pinia, router],
-      },
-    });
-  
-    await wrapper.vm.$nextTick();
-  
-    const containerClass = (wrapper.vm as any).containerClass;
-  
-    expect(containerClass['layout-theme-light']).toBe(true);
-    expect(containerClass['layout-static']).toBe(true);
-    expect(containerClass['layout-overlay']).toBe(false);
+    const topbar = baseLayout.findComponent(AppTopbar);
+    expect(topbar.exists()).toBe(true);
+
+    const sidebar = baseLayout.findComponent(ConsultantMenu);
+    expect(sidebar.exists()).toBe(true);
+
+    const footer = baseLayout.findComponent(AppFooter);
+    expect(footer.exists()).toBe(true);
   });
 });
